refactor(config-ui): adjust the type for create bp page
