+++ config-ui/src/pages/blueprint/index.ts
@@ -16,5 +16,6 @@
  *
  */
 
+export * from './types'
 export * from './create'
 export * from './detail'
